requirements.txt
opencv-python-headless
streamlit
tensorflow
scikit-learn
joblib
numpy
Pillow
gdown